requirements.txt
fastapi==0.95.0
uvicorn==0.23.0
torch==2.1.0
torchvision==0.15.2
pillow==9.4.0